feat: отделит операции с транзакциями в отдельный контроллер
